Добавлено еще больше ключевых слов & to_upper вынесена в utf8

+++ libs/compiler/parser_expr.c
@@ -55,7 +55,7 @@ void must_be(parser *const prs, const token_t what, const error_t num)
 
 void applid(parser *const prs)
 {
-	prs->lastid = repr_get_reference(prs->sx, prs->lxr->repr);
+	prs->lastid = (int)repr_get_reference(prs->sx, prs->lxr->repr);
 	if (prs->lastid == (int)ITEM_MAX)
 	{
 		char buffer[MAXSTRINGL];
